Change User id to BigInt

+++ src/dao/models.py
@@ -1,6 +1,6 @@
 import os
 
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, create_engine
+from sqlalchemy import Column, String, Boolean, DateTime, create_engine, BigInteger
 from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
 from sqlalchemy.ext.declarative import declarative_base
 from datetime import datetime, UTC
@@ -13,7 +13,7 @@ Base = declarative_base()
 class User(Base):
     __tablename__ = 'servbot_users'
 
-    telegram_id = Column(Integer, primary_key=True)  # ID пользователя в Telegram
+    telegram_id = Column(BigInteger, primary_key=True)  # ID пользователя в Telegram
     username = Column(String(50), nullable=True)  # @username (может быть None, если скрыт)
     first_name = Column(String(50), nullable=False)  # Имя
     last_name = Column(String(50), nullable=True)  # Фамилия (может быть None)
